refactor: 클래스 이름 변경, 패키지 이동
