+++ src/main/java/racingcar/service/GameService.java
@@ -1,4 +1,4 @@
-package racingcar.controller;
+package racingcar.service;
 
 import racingcar.domain.Car;
 import racingcar.view.Output;
@@ -6,7 +6,7 @@ import racingcar.view.Output;
 import java.util.ArrayList;
 import java.util.List;
 
-public class CarController {
+public class GameService {
 
     public static void raceStart(List<Car> cars, int tryNum) {
         Output.printStartRace();
@@ -17,9 +17,9 @@ public class CarController {
     }
 
     private static void move(List<Car> cars) {
-        for (Car car : cars) {
-            car.goForward();
-        }
+//        for (Car car : cars) {
+//            car.goForward();
+//        }
     }
 
     public static void finalWinner(List<Car> cars) {
@@ -28,13 +28,13 @@ public class CarController {
 
     private static List<String> getWinnerList(List<Car> cars) {
         List<String> winnerList = new ArrayList<>();
-        int winnerPosition = Car.getWinnerPosition(cars);
+//        int winnerPosition = Car.getWinnerPosition(cars);
 
-        for (Car car : cars) {
-            if (car.isWinner(winnerPosition)) {
-                winnerList.add(car.getName());
-            }
-        }
+//        for (Car car : cars) {
+//            if (car.isWinner(winnerPosition)) {
+//                winnerList.add(car.getName());
+//            }
+//        }
 
         return winnerList;
     }
