+++ src/main/java/racingcar/util/Validation.java
@@ -1,4 +1,4 @@
-package racingcar.controller;
+package racingcar.util;
 
 import java.util.Arrays;
 import java.util.List;
